core/pyproject.toml
[project]
name = "ttsdb-core"
version = "0.1.4"
description = "Core package for TTSDB systems - base classes for voice cloning TTS models."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    # torch intentionally omitted - install separately via:
    #   CPU: uv pip install torch --index-url https://download.pytorch.org/whl/cpu
    #   GPU: uv pip install torch --index-url https://download.pytorch.org/whl/cu121
    "transformers>=4.20.0",
    "numpy>=1.21.0",
    "soundfile>=0.10.0",
    "scipy>=1.7.0",
    "pyyaml>=6.0",
    "importlib_resources>=5.0; python_version < '3.9'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ttsdb_core"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""
